requirements/prod.txt
annotated-types==0.6.0 ; python_version >= "3.9" and python_version < "4.0"
argparse==1.4.0 ; python_version >= "3.9" and python_version < "4.0"
asn1crypto==1.5.1 ; python_version >= "3.9" and python_version < "4.0"
attrs==23.1.0 ; python_version >= "3.9" and python_version < "4.0"
cffi==1.16.0 ; python_version >= "3.9" and python_version < "4.0"
cryptography==41.0.4 ; python_version >= "3.9" and python_version < "4.0"
dlms-cosem==21.3.2 ; python_version >= "3.9" and python_version < "4.0"
dsmr-parser==1.3.0 ; python_version >= "3.9" and python_version < "4.0"
paho-mqtt==1.6.1 ; python_version >= "3.9" and python_version < "4.0"
psycopg2-binary==2.9.9 ; python_version >= "3.9" and python_version < "4.0"
pycparser==2.21 ; python_version >= "3.9" and python_version < "4.0"
pydantic-core==2.10.1 ; python_version >= "3.9" and python_version < "4.0"
pydantic-settings==2.0.3 ; python_version >= "3.9" and python_version < "4.0"
pydantic==2.4.2 ; python_version >= "3.9" and python_version < "4.0"
pyserial-asyncio==0.6 ; python_version >= "3.9" and python_version < "4.0"
pyserial==3.5 ; python_version >= "3.9" and python_version < "4.0"
python-dateutil==2.8.2 ; python_version >= "3.9" and python_version < "4.0"
python-dotenv==1.0.0 ; python_version >= "3.9" and python_version < "4.0"
pytz==2023.3.post1 ; python_version >= "3.9" and python_version < "4.0"
pyyaml==6.0.1 ; python_version >= "3.9" and python_version < "4.0"
six==1.16.0 ; python_version >= "3.9" and python_version < "4.0"
tailer==0.4.1 ; python_version >= "3.9" and python_version < "4.0"
typing-extensions==4.8.0 ; python_version >= "3.9" and python_version < "4.0"
